+++ paddle/fluid/operators/controlflow/tensor_array_read_write_op.cc
@@ -169,7 +169,7 @@ class ReadFromArrayOp : public ArrayOp {
       auto &fw_var_tensor = fw_var->Get<framework::LoDTensor>();
 
       framework::AttributeMap attrs;
-      attrs["dtype"] = fw_var_tensor.type();
+      attrs["dtype"] = static_cast<int>(fw_var_tensor.type());
       attrs["shape"] = framework::vectorize<int>(fw_var_tensor.dims());
       attrs["value"] = 0.0f;
