+++ paddle/fluid/operators/rnn_memory_helper_op.cc
@@ -116,7 +116,7 @@ class RNNMemoryHelperGradOp : public framework::OperatorBase {
       auto &in_var_tensor = in_var->Get<framework::LoDTensor>();
 
       framework::AttributeMap attrs;
-      attrs["dtype"] = in_var_tensor.type();
+      attrs["dtype"] = static_cast<int>(in_var_tensor.type());
       attrs["shape"] = framework::vectorize<int>(in_var_tensor.dims());
       attrs["value"] = 0.0f;
